# starlark: fall back to slow Ints if mmap fails