@adonovan 2022-08-16
> Seems reasonable. Would be good to have a test, though (e.g. set RLIMIT_AS, see if int tests still pass).

Done, though it's a pain because it requires fork+execing a child process, and it requires ulimit -v which is not available on MacOS or Windows.