@adonovan 2022-08-16
@tetromino, PTAL at the changes to the build tags, go.mod/sum files, Actions workflow config, and of course the test.